requirements.txt
streamlit>=1.28.0
fastapi>=0.104.1
sqlalchemy>=2.0.23
authlib>=1.2.1
python-jose>=3.3.0
python-multipart>=0.0.6
pymupdf>=1.23.5
unstructured>=0.10.30
langchain>=0.0.330
langchain-community>=0.0.10
openai>=1.3.0
pinecone-client>=2.2.4
boto3>=1.34.0
python-dotenv>=1.0.0
uvicorn>=0.24.0
psycopg2-binary>=2.9.9
pydantic>=2.5.1
bcrypt>=4.0.1
passlib>=1.7.4
docling>=0.1.0
tiktoken>=0.5.2
transformers>=4.37.2
requests>=2.31.0
humanize>=4.9.0
asyncpg==0.29.0
aiosqlite==0.19.0
